start moving to pybind11, and delete swig_matlab bindings
