+++ iris/CMakeLists.txt
@@ -4,23 +4,17 @@ if(APPLE AND POLICY CMP0042)
   cmake_policy(SET CMP0042 NEW)
 endif()
 
-set(CMAKE_MODULE_PATH ${CMAKE_CURRENT_SOURCE_DIR}/cmake ${CMAKE_CURRENT_SOURCE_DIR}/swigmake/cmake ${CMAKE_MODULE_PATH} )
-
-if(CMAKE_VERSION VERSION_LESS 3.7)
-  set(CMAKE_MODULE_PATH "${CMAKE_CURRENT_SOURCE_DIR}/cmake/3.7" ${CMAKE_MODULE_PATH})
-endif()
-
 include(CTest)
+find_package(pybind11 REQUIRED)
 
-# pull in the pods macros. See cmake/pods.cmake for documentation
-set(POD_NAME iris)
-include(pods)
+# Set RPATH for installed binaries
+set(CMAKE_INSTALL_RPATH
+  ${CMAKE_INSTALL_PREFIX}/${CMAKE_INSTALL_LIBDIR}
+  ${CMAKE_INSTALL_PREFIX}/lib${LIB_SUFFIX}
+  ${CMAKE_INSTALL_PREFIX}/lib)
+set(CMAKE_INSTALL_RPATH_USE_LINK_PATH ON)
 
-find_package(Matlab COMPONENTS MEX_COMPILER MX_LIBRARY)
-
-if(Matlab_FOUND)
-  include(matlab_pods)
-endif()
+list(REMOVE_DUPLICATES CMAKE_INSTALL_RPATH)
 
 if (CMAKE_COMPILER_IS_GNUCC)
   set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wreturn-type -Wuninitialized -Wunused-variable") # -Wunused-parameter")
@@ -46,23 +40,14 @@ elseif (MSVC)
 endif()
 
 # install all python files in the python/ subdirectory
-pods_install_python_packages(${CMAKE_CURRENT_SOURCE_DIR}/python)
+# pods_install_python_packages(${CMAKE_CURRENT_SOURCE_DIR}/python)
 
 set(CMAKE_LIBRARY_PATH ${CMAKE_LIBRARY_PATH} ${CMAKE_INSTALL_PREFIX}/lib)
 
 find_package(Eigen3 REQUIRED)
 include_directories( ${EIGEN3_INCLUDE_DIR} )
 
-add_subdirectory(cvxgen)
 add_subdirectory(src)
 
-if(Matlab_FOUND)
-  add_subdirectory(matlab/+iris)
-
-  # setup matlab pods-compliance
-  pods_configure_matlab_paths()
-  pods_install_matlab_path(iris)
-endif()
-
 add_subdirectory(python)
-pods_install_bash_setup(iris "export PYTHONPATH=${CMAKE_INSTALL_PREFIX}/lib/python2.7/dist-packages:\$PYTHONPATH")
+# pods_install_bash_setup(iris "export PYTHONPATH=${CMAKE_INSTALL_PREFIX}/lib/python2.7/dist-packages:\$PYTHONPATH")
