+++ iris/python/CMakeLists.txt
@@ -1,16 +1,19 @@
-include(SwigPython)
+install(DIRECTORY irispy DESTINATION lib/python${PYTHON_VERSION_MAJOR}.${PYTHON_VERSION_MINOR}/dist-packages)
 
-find_package(PythonInterp REQUIRED)
-set(python_install_dir lib/python${PYTHON_VERSION_MAJOR}.${PYTHON_VERSION_MINOR}/dist-packages)
 
-add_swig_python_module(irisWrapperPython iris_wrapper.i
-                       CPLUSPLUS
-                       SWIG_INCLUDE_DIRS ${CMAKE_CURRENT_SOURCE_DIR}/../swigmake/swig/python
-                       INCLUDE_DIRS ${CMAKE_CURRENT_SOURCE_DIR}/../src
-                       LINK_LIBRARIES iris
-                       DESTINATION ${python_install_dir}/irispy ${CMAKE_CURRENT_SOURCE_DIR}/irispy)
+# include(SwigPython)
 
-pods_install_python_packages(${CMAKE_CURRENT_SOURCE_DIR})
+# find_package(PythonInterp REQUIRED)
+# set(python_install_dir lib/python${PYTHON_VERSION_MAJOR}.${PYTHON_VERSION_MINOR}/dist-packages)
+
+# add_swig_python_module(irisWrapperPython iris_wrapper.i
+#                        CPLUSPLUS
+#                        SWIG_INCLUDE_DIRS ${CMAKE_CURRENT_SOURCE_DIR}/../swigmake/swig/python
+#                        INCLUDE_DIRS ${CMAKE_CURRENT_SOURCE_DIR}/../src
+#                        LINK_LIBRARIES iris
+#                        DESTINATION ${python_install_dir}/irispy ${CMAKE_CURRENT_SOURCE_DIR}/irispy)
+
+# pods_install_python_packages(${CMAKE_CURRENT_SOURCE_DIR})
 
 if(BUILD_TESTING)
   add_test(NAME python_tests
