+++ iris/python/irispy/__init__.py
@@ -1,1 +1,3 @@
-from irispy import inflate_region, Polyhedron, Ellipsoid, IRISRegion, IRISOptions, IRISProblem
+from __future__ import absolute_import
+
+from .irispy import inflate_region, Polyhedron, Ellipsoid, IRISRegion, IRISOptions, IRISProblem
