+++ iris/src/cvxgen/CMakeLists.txt
@@ -1,8 +1,10 @@
 add_library(iris_cvxgen_ldp SHARED solver.c matrix_support.c ldl.c util.c cvxgen_ldp.c)
-pods_install_libraries(iris_cvxgen_ldp)
-pods_install_headers(cvxgen_ldp.h DESTINATION iris)
+# pods_install_headers(cvxgen_ldp.h DESTINATION iris)
 
 add_library(iris_cvxgen_ldp_cpp SHARED cvxgen_ldp.cpp)
 target_link_libraries(iris_cvxgen_ldp_cpp iris_cvxgen_ldp)
-pods_install_libraries(iris_cvxgen_ldp_cpp)
-pods_install_headers(cvxgen_ldp.h DESTINATION iris)
+
+install(TARGETS iris_cvxgen_ldp_cpp iris_cvxgen_ldp
+        LIBRARY DESTINATION lib)
+
+# pods_install_headers(cvxgen_ldp.h DESTINATION iris)
